requirements.txt
numpy==1.20.2
pandas==1.3.5
Pillow==9.2.0
scikit_learn==1.1.2
streamlit==1.12.2
